requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cross-web==0.4.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.129.0
graphql-core==3.2.7
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
strawberry-graphql==0.294.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0

# Test dependencies
pytest>=9.0.0
httpx>=0.28.0
